pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "climbing-project"
version = "0.1.0"
description = "Raspberry Pi + ESP32 laser projection system for climbing routes"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask",
    "opencv-python",
    "numpy",
    "pyserial",
    "werkzeug",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "flake8",
    "mypy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["vision*", "web_interface*"]
